GOCBC-930: Add support for CreateAsDeleted
Motivation
----------
Transactions are going to require support for a new MutateIn
flag: CreateAsDeleted.

Changes
-------
Add HelloFeature for CreateAsDeleted.
Add support to kv mux state to extract server support from the
cluster config.
Add option to MutateInOptions.
Wire up the option.
Add IsDeleted field to LookupInResult as Internal.

Reviewed-on: http://review.couchbase.org/c/gocbcore/+/131371
Tested-by: Charles Dixon <chvckd@gmail.com>
